+++ public/app/flutter_bootstrap.js
@@ -11,6 +11,6 @@ _flutter.buildConfig = {"engineRevision":"a18df97ca57a249df5d8d68cd0820600223ce2
 
 _flutter.loader.load({
   serviceWorkerSettings: {
-    serviceWorkerVersion: "2626028985"
+    serviceWorkerVersion: "3933159691"
   }
 });